+++ js/core/Ragdoll3DDesktop.ts
@@ -18,9 +18,9 @@ export class Ragdoll3DDesktop extends Ragdoll3DCore implements IRagdoll3DControl
         
         this.audioManager = Services.get('AudioManager') || null;
         if (this.audioManager) {
-            void this.audioManager.loadSound('scream', '/games/ragdoll/assets/audio/ahh.opus');
-            void this.audioManager.loadSound('wii', '/games/ragdoll/assets/audio/wii.opus');
-            void this.audioManager.loadSound('boing', '/games/ragdoll/assets/audio/boing.opus');
+            void this.audioManager.loadSound('scream', Utils.getAssetUrl('/games/ragdoll/assets/audio/ahh.opus'));
+            void this.audioManager.loadSound('wii', Utils.getAssetUrl('/games/ragdoll/assets/audio/wii.opus'));
+            void this.audioManager.loadSound('boing', Utils.getAssetUrl('/games/ragdoll/assets/audio/boing.opus'));
         }
         
         const BubbleAnimatorCtor = Services.get('BubbleAnimator');